+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Support async validation within the `optional` rule ([#206](https://github.com/imbrn/v8n/issues/206))
 
+### Fixed
+
+- Correct incorrect return type of `check` validation strategy (boolean -> void)
+
 ## [1.4.0] - 2022-02-17
 
 ### Added