+++ CHANGELOG.md
@@ -9,8 +9,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- New rule `optionalAsync` for asynchronous, optional validation ([#206](https://github.com/imbrn/v8n/issues/206))
+- Support async validation within the `optional` rule ([#206](https://github.com/imbrn/v8n/issues/206))
 - New modifier `strict` for schema validation ([#191](https://github.com/imbrn/v8n/issues/191), [#179](https://github.com/imbrn/v8n/issues/179))
+
+### Fixed
+
+- Correct incorrect return type of `check` validation strategy (boolean -> void)
 
 ## [1.4.0] - 2022-02-17