+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.5.0] - 2022-06-27
+
 ### Added
 
 - Support async validation within the `optional` rule ([#206](https://github.com/imbrn/v8n/issues/206))
@@ -127,7 +129,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Build process now properly transpiles modules from ES6 to ES5. ([#44](https://github.com/imbrn/v8n/issues/44))
 
-[unreleased]: https://github.com/imbrn/v8n/compare/v1.4.0...HEAD
+[unreleased]: https://github.com/imbrn/v8n/compare/v1.5.0...HEAD
+[1.5.0]: https://github.com/imbrn/v8n/compare/v1.4.0...v1.5.0
 [1.4.0]: https://github.com/imbrn/v8n/compare/v1.3.3...v1.4.0
 [1.3.3]: https://github.com/imbrn/v8n/compare/v1.3.2...v1.3.3
 [1.3.2]: https://github.com/imbrn/v8n/compare/v1.3.1...v1.3.2